+++ src/main.rs
@@ -15,12 +15,9 @@ extern crate pipeline;
 use clap::{crate_version, App, Arg};
 use config::RootConfig;
 use discord::Handler;
-use serenity::{
-    client::{bridge::gateway::GatewayIntents, validate_token},
-    prelude::*,
-};
-use std::{env, error::Error, process, sync::Arc};
-use tracing::{info, warn, Level};
+use serenity::{model::gateway::GatewayIntents, prelude::*};
+use std::{env, error::Error, sync::Arc};
+use tracing::{info, Level};
 
 struct ConfigContainer;
 
@@ -87,18 +84,23 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     // validate the bot token
     let bot_token = env::var("DISCORD_TOKEN").expect("expected a Discord token in the environment");
-    if validate_token(bot_token.clone()).is_err() {
-        warn!("+--------------------------------------------------------------------------------------------+");
-        warn!("| Discord bot token is either missing or invalid!                                            |");
-        warn!("|                                                                                            |");
-        warn!("| Create a Discord bot here:                                                                 |");
-        warn!("| https://discord.com/developers/applications/me                                             |");
-        warn!("|                                                                                            |");
-        warn!("| Copy the token into your config file, and add the bot to your server with this URL:        |");
-        warn!("| https://discord.com/oauth2/authorize?client_id=<BOT CLIENT ID>&permissions=10240&scope=bot |");
-        warn!("+--------------------------------------------------------------------------------------------+");
-        process::exit(0);
-    }
+    // TODO: Token validation is currently broken in Serenity. Check back later
+    // match validate_token(bot_token.clone()) {
+    //     Ok(()) => (),
+    //     Err(e) => {
+    //         warn!("+--------------------------------------------------------------------------------------------+");
+    //         warn!("| Discord bot token is either missing or invalid!                                            |");
+    //         warn!("| Error: {}", e);
+    //         warn!("|                                                                                            |");
+    //         warn!("| Create a Discord bot here:                                                                 |");
+    //         warn!("| https://discord.com/developers/applications/me                                             |");
+    //         warn!("|                                                                                            |");
+    //         warn!("| Copy the token into your config file, and add the bot to your server with this URL:        |");
+    //         warn!("| https://discord.com/oauth2/authorize?client_id=<BOT CLIENT ID>&permissions=10240&scope=bot |");
+    //         warn!("+--------------------------------------------------------------------------------------------+");
+    //         process::exit(0);
+    //     }
+    // }
 
     let application_id: u64 = env::var("DISCORD_APPLICATION_ID")
         .expect("Expect a Discord application ID in the environment")
@@ -108,17 +110,18 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let handler = Handler::new(Arc::clone(&cfg_lock));
 
     // Create the Discord client
-    let mut client = Client::builder(&bot_token)
-        .application_id(application_id)
-        .event_handler(handler)
-        .intents(
-            GatewayIntents::GUILDS
-                | GatewayIntents::GUILD_MEMBERS
-                | GatewayIntents::GUILD_MESSAGES
-                | GatewayIntents::GUILD_PRESENCES,
-        )
-        .await
-        .expect("Error creating Discord client");
+    let mut client = Client::builder(
+        &bot_token,
+        GatewayIntents::GUILDS
+            | GatewayIntents::GUILD_MEMBERS
+            | GatewayIntents::GUILD_MESSAGES
+            | GatewayIntents::GUILD_PRESENCES
+            | GatewayIntents::MESSAGE_CONTENT,
+    )
+    .application_id(application_id)
+    .event_handler(handler)
+    .await
+    .expect("Error creating Discord client");
 
     {
         let mut data = client.data.write().await;